pyproject.toml
[tool.poetry]
name = "slackbot-flask"
version = "0.1.0"
description = ""
authors = ["epeltz33 <epeltz@optonline.net>"]
readme = "README.md"
packages = [{include = "slackbot_flask"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
slackclient = "^2.9.4"
slackeventsapi = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
